+++ tests/integration/mod.rs
@@ -42,7 +42,6 @@ pub const JWT_SECRET: &str = "688f5d737bad920bdfb2fc2f488d6b6209eebda1dae949a8de
 pub const L2_P2P_ENODE: &str = "3479db4d9217fb5d7a8ed4d61ac36e120b05d36c2eefb795dc42ff2e971f251a2315f5649ea1833271e020b9adc98d5db9973c7ed92d6b2f1f2223088c3d852f";
 
 mod containers;
-mod integration_test;
 mod proxy;
 
 pub struct LoggingConsumer {
